Move AntlrParserResult cache to AntlrParser and update that regularly.
