+++ java/languages.antlr/src/org/netbeans/modules/languages/antlr/AntlrParser.java
@@ -18,11 +18,21 @@
  */
 package org.netbeans.modules.languages.antlr;
 
+import java.lang.ref.Reference;
+import java.lang.ref.SoftReference;
+import java.util.Collections;
+import java.util.WeakHashMap;
 import javax.swing.event.ChangeListener;
+import org.netbeans.modules.parsing.api.ParserManager;
+import org.netbeans.modules.parsing.api.ResultIterator;
 import org.netbeans.modules.parsing.api.Snapshot;
+import org.netbeans.modules.parsing.api.Source;
 import org.netbeans.modules.parsing.api.Task;
+import org.netbeans.modules.parsing.api.UserTask;
 import org.netbeans.modules.parsing.spi.ParseException;
 import org.netbeans.modules.parsing.spi.SourceModificationEvent;
+import org.openide.filesystems.FileObject;
+import org.openide.util.Exceptions;
 
 /**
  *
@@ -30,8 +40,16 @@ import org.netbeans.modules.parsing.spi.SourceModificationEvent;
  */
 public abstract class AntlrParser extends org.netbeans.modules.parsing.spi.Parser {
 
-
     AntlrParserResult lastResult;
+
+    /*
+     * The NetBeans indexer support is more appropriate for large scale project
+     * where parsing of all affected files would take a too long time. For ANTLR
+     * imports are only possible from the current directory, so it can be
+     * expected that the number of files that can be imported or need to be
+     * scanned is low.
+     */
+    private static final WeakHashMap<FileObject, Reference<AntlrParserResult>> CACHE = new WeakHashMap<>();
 
     protected abstract AntlrParserResult<?> createParserResult(Snapshot snapshot);
 
@@ -39,7 +57,9 @@ public abstract class AntlrParser extends org.netbeans.modules.parsing.spi.Parse
     public void parse(Snapshot snapshot, Task task, SourceModificationEvent event) throws ParseException {
         AntlrParserResult<?> parserResult = createParserResult(snapshot);
 
-        lastResult = parserResult.get();
+        AntlrParserResult<?> parsed = parserResult.get();
+        cacheResult(snapshot.getSource().getFileObject(), parsed);
+        lastResult = parsed;
     }
 
     @Override
@@ -54,6 +74,45 @@ public abstract class AntlrParser extends org.netbeans.modules.parsing.spi.Parse
 
     @Override
     public void removeChangeListener(ChangeListener changeListener) {
+    }
+
+    private static void cacheResult(FileObject fo, AntlrParserResult<?> result) {
+        synchronized (CACHE) {
+            CACHE.put(fo, new SoftReference<>(result));
+        }
+    }
+
+    public static AntlrParserResult getParserResult(FileObject fo) {
+        AntlrParserResult result = null;
+        java.lang.ref.Reference<AntlrParserResult> ceReference;
+        synchronized (CACHE) {
+            ceReference = CACHE.get(fo);
+        }
+        if (ceReference != null) {
+            result = ceReference.get();
+        }
+
+        if (result == null) {
+            try {
+                AntlrParserResult[] parserResult = new AntlrParserResult[1];
+                ParserManager.parse(Collections.singleton(Source.create(fo)), new UserTask() {
+                    @Override
+                    public void run(ResultIterator resultIterator) throws Exception {
+                        org.netbeans.modules.parsing.spi.Parser.Result result = resultIterator.getParserResult();
+                        if (result instanceof AntlrParserResult) {
+                            parserResult[0] = (AntlrParserResult) result;
+                        }
+                    }
+                });
+                if (parserResult[0] != null) {
+                    result = parserResult[0];
+                }
+            } catch (ParseException ex) {
+                Exceptions.printStackTrace(ex);
+            }
+        }
+
+        return result;
     }
 
 }
